render markdown & handle uncompletable tasks
